HDFS-15989. Split TestBalancer and De-flake testMaxIterationTime()
